Split TrailingComma into TrailingCommaInLiteral and TrailingCommaInArguments
This will allow users to disable one or the other, or set different trailing
comma styles for literals and method calls.
