+++ lib/rubocop.rb
@@ -77,6 +77,7 @@ require 'rubocop/cop/mixin/space_inside' # relies on surrounding_space
 require 'rubocop/cop/mixin/statement_modifier'
 require 'rubocop/cop/mixin/string_help'
 require 'rubocop/cop/mixin/string_literals_help'
+require 'rubocop/cop/mixin/trailing_comma'
 require 'rubocop/cop/mixin/unused_argument'
 
 require 'rubocop/cop/lint/ambiguous_operator'
@@ -324,7 +325,8 @@ require 'rubocop/cop/style/symbol_literal'
 require 'rubocop/cop/style/symbol_proc'
 require 'rubocop/cop/style/tab'
 require 'rubocop/cop/style/trailing_blank_lines'
-require 'rubocop/cop/style/trailing_comma'
+require 'rubocop/cop/style/trailing_comma_in_arguments'
+require 'rubocop/cop/style/trailing_comma_in_literal'
 require 'rubocop/cop/style/trailing_underscore_variable'
 require 'rubocop/cop/style/trailing_whitespace'
 require 'rubocop/cop/style/trivial_accessors'
